fix: move fixtures to conftest.py
